refactor: replace deprecated ioutil.ReadAll with io.ReadAll
- Update import from io/ioutil to io package
- Replace ioutil.ReadAll() call with io.ReadAll()
- Align variable declarations for better readability

This change follows Go 1.16+ recommendations for handling io operations.
